fix: keep Cycle 873 acceptance source transportable
