+++ scripts/frontier_cycle873_recurrent_f17_uniform_affine_open_box_acceptance_2026_08_03.py
@@ -22,6 +22,10 @@ import sys
 import tempfile
 from typing import TYPE_CHECKING
 
+from frontier_cycle873_source_view_validation_2026_08_03 import (
+    source_view_certificate,
+)
+
 
 # These imports are intentionally type-check-only.  Each module is an
 # import-free, byte-exact source view, so the restricted audit packet receives
@@ -54,6 +58,7 @@ ACCEPTANCE = "scripts/frontier_cycle873_recurrent_f17_uniform_affine_open_box_ac
 PHYSICAL = "scripts/frontier_cycle873_recurrent_f17_all_seam_physical_core_2026_08_03.py"
 LOCAL = "scripts/frontier_cycle873_f17_open_box_local_constraints_core_2026_08_03.py"
 AFFINE = "scripts/frontier_cycle873_uniform_affine_gauss_intertwiner_core_2026_08_03.py"
+SOURCE_VIEW_VALIDATION = "scripts/frontier_cycle873_source_view_validation_2026_08_03.py"
 PRIMARY_RECEIPT = "outputs/cycle873_recurrent_f17_uniform_affine_open_box_primary_receipt_2026_08_03.json"
 INDEPENDENT_RECEIPT = "outputs/cycle873_recurrent_f17_uniform_affine_open_box_independent_check_receipt_2026_08_03.json"
 PHYSICAL_RECEIPT = "outputs/cycle873_recurrent_f17_all_seam_physical_core_receipt_2026_08_03.json"
@@ -65,7 +70,7 @@ NOTE = "docs/RECURRENT_F17_UNIFORM_AFFINE_OPEN_BOX_CYCLE873_BOUNDED_THEOREM_NOTE
 PRIMARY_LOG = "logs/runner-cache/frontier_cycle873_recurrent_f17_uniform_affine_open_box_primary_2026_08_03.txt"
 INDEPENDENT_LOG = "logs/runner-cache/frontier_cycle873_recurrent_f17_uniform_affine_open_box_independent_check_2026_08_03.txt"
 DEFAULT_OUTPUT = ROOT / ACCEPTANCE_RECEIPT
-EXPECTED_MANIFEST_SHA256 = "f5239fc8436f39c4043e1028968d79407934d9c9b9464cbcb415d1700db8591d"
+EXPECTED_MANIFEST_SHA256 = "4382c9eb7c73ef325a48e85c1c35ee474f8eb0438e0283a018bb62020c176774"
 
 LOCAL_SOURCE_VIEWS = (
     "scripts/frontier_cycle873_local_constraints_source_audit_view_part1_2026_08_03.py",
@@ -120,6 +125,7 @@ AUDIT_INPUT_PATHS = (
     "scripts/frontier_cycle873_recurrent_f17_uniform_affine_open_box_independent_check_2026_08_03.py",
     "scripts/frontier_cycle873_recurrent_f17_uniform_affine_open_box_primary_2026_08_03.py",
     "scripts/frontier_cycle873_uniform_affine_gauss_intertwiner_core_2026_08_03.py",
+    "scripts/frontier_cycle873_source_view_validation_2026_08_03.py",
     "scripts/frontier_cycle873_local_constraints_source_audit_view_part1_2026_08_03.py",
     "scripts/frontier_cycle873_local_constraints_source_audit_view_part2_2026_08_03.py",
     "scripts/frontier_cycle873_local_constraints_source_audit_view_part3_2026_08_03.py",
@@ -140,7 +146,7 @@ AUDIT_INPUT_PATHS = (
 DECLARED_INPUT_PATHS = AUDIT_INPUT_PATHS
 
 EXPECTED_ARTIFACT_SHA256 = {
-    NOTE: "3cd18032e6f14007e11c2709e70021b9fbcd81fcfc8974cbc650f64cd4df010e",
+    NOTE: "37bf888a461aac7f3929d0126a7a2b4c21f4fd01cc8e299ff6f8b42eba88e67f",
     PRIMARY_LOG: "72be1cd183bf85663c6726223f9db9fb61be3e80b6ed0f3d386f32849312eb74",
     INDEPENDENT_LOG: "6689ee8aac078f36507af89c20a90a19ea0721d34af74b1a98318cec06c35812",
     PHYSICAL_RECEIPT: "397657af570393fad9967edc55e74f7a66f46e8284fd5102be0f5e1df9247d0b",
@@ -153,6 +159,7 @@ EXPECTED_ARTIFACT_SHA256 = {
     AFFINE: "a1bc2159c5e2d5f59087860e3fe40bb1919cd4e476f6565a99c326d5af1c5ca9",
     PRIMARY: "ab9f365c167b8fafb4f54508c0fb38b325bf687fdf8f222bc9aa833ad65dfc62",
     INDEPENDENT: "02c3f321ba5ef1dce723ed04bd83919839648fd89202f607b6cc680645a97734",
+    SOURCE_VIEW_VALIDATION: "4b21203146df10b4a85ddcc245aa6a62f5cb03de8873756ccd745e4ba809b2c4",
     LOCAL_SOURCE_VIEWS[0]: "b85d277b1fc6e4a8fabd1e62c52d369bbe8d2f5762ce7f44eac35539c09ff16e",
     LOCAL_SOURCE_VIEWS[1]: "24441e0224609c66bfd3cb3e7e924e2cef071125d1f90f8ca26b84487074f779",
     LOCAL_SOURCE_VIEWS[2]: "207f75b20961a5c33d40dc2a38fa26c753ce3287d5655f7bb6815b35ddae2aeb",
@@ -188,6 +195,7 @@ EXPECTED_PACKAGE_FILES = tuple(sorted((
     PRIMARY,
     INDEPENDENT,
     ACCEPTANCE,
+    SOURCE_VIEW_VALIDATION,
     *SOURCE_VIEW_FILES,
 )))
 
@@ -225,7 +233,7 @@ NOTE_REQUIRED_TEXT = (
     "supplied lattice parity origin",
     "does not prove unit-translation/origin-shift equivalence",
     "The 20 M2 persistent bank is not the whole routing substrate.",
-    "The exact candidate surface has 32 files:",
+    "The exact candidate surface has 33 files:",
     "16 import-free byte-exact source-view modules",
     "[Minimal framework axioms](MINIMAL_AXIOMS_2026-06-29.md)",
     "[Cycle 870 recurrent physical-M2 matter compiler]",
@@ -238,173 +246,6 @@ def digest(path: Path) -> str:
 
 def load_json(path: Path) -> dict:
     return json.loads(path.read_text(encoding="utf-8"))
-
-
-def literal_assignment_bytes(source: bytes, name: str):
-    tree = ast.parse(source.decode("utf-8", errors="strict"))
-    for node in tree.body:
-        if not isinstance(node, (ast.Assign, ast.AnnAssign)):
-            continue
-        targets = node.targets if isinstance(node, ast.Assign) else (node.target,)
-        if any(
-            isinstance(target, ast.Name) and target.id == name
-            for target in targets
-        ):
-            return ast.literal_eval(node.value)
-    raise KeyError(name)
-
-
-def parse_source_view(raw: bytes, label: str) -> dict[str, object]:
-    metadata = {
-        key: literal_assignment_bytes(raw, key)
-        for key in (
-            "TARGET_SOURCE", "PART_ORDINAL", "PART_COUNT", "FIRST_SOURCE_LINE",
-            "LAST_SOURCE_LINE", "TOTAL_SOURCE_LINES", "SOURCE_FINAL_NEWLINE",
-            "EXPECTED_SOURCE_SHA256",
-        )
-    }
-    prefix = b"# C873SRC "
-    rows: list[tuple[int, bytes]] = []
-    for line in raw.splitlines(keepends=True):
-        if line.startswith(prefix):
-            if not line.endswith(b"\n") or line.endswith(b"\r\n"):
-                raise ValueError(f"{label}: payload newline")
-            number_raw, separator, payload = line[len(prefix):].partition(b"|")
-            if separator != b"|" or len(number_raw) != 6 or not number_raw.isdigit():
-                raise ValueError(f"{label}: payload prefix")
-            rows.append((int(number_raw), payload[:-1]))
-        elif line.startswith(b"# C873SRC"):
-            raise ValueError(f"{label}: malformed payload prefix")
-    first = int(metadata["FIRST_SOURCE_LINE"])
-    last = int(metadata["LAST_SOURCE_LINE"])
-    if [number for number, _ in rows] != list(range(first, last + 1)):
-        raise ValueError(f"{label}: noncontiguous payload rows")
-    metadata["rows"] = rows
-    metadata["view_path"] = label
-    return metadata
-
-
-def validate_source_view_set(
-    target: str,
-    view_paths: tuple[str, ...],
-    *,
-    raw_overrides: dict[str, bytes] | None = None,
-    actual_override: bytes | None = None,
-    enforce_view_hashes: bool = True,
-) -> dict[str, object]:
-    raw_overrides = raw_overrides or {}
-    parts = []
-    expected_next = 1
-    total_lines = None
-    final_newline = None
-    expected_source_sha = EXPECTED_ARTIFACT_SHA256[target]
-    for ordinal, view_path in enumerate(view_paths, 1):
-        raw = raw_overrides.get(view_path, (ROOT / view_path).read_bytes())
-        if (
-            enforce_view_hashes
-            and sha256(raw).hexdigest() != EXPECTED_ARTIFACT_SHA256[view_path]
-        ):
-            raise ValueError(f"{view_path}: view hash")
-        part = parse_source_view(raw, view_path)
-        if part["TARGET_SOURCE"] != target:
-            raise ValueError(f"{view_path}: target")
-        if part["PART_ORDINAL"] != ordinal or part["PART_COUNT"] != len(view_paths):
-            raise ValueError(f"{view_path}: ordinal")
-        if part["FIRST_SOURCE_LINE"] != expected_next:
-            raise ValueError(f"{view_path}: gap or overlap")
-        if part["EXPECTED_SOURCE_SHA256"] != expected_source_sha:
-            raise ValueError(f"{view_path}: source hash declaration")
-        if total_lines is None:
-            total_lines = part["TOTAL_SOURCE_LINES"]
-            final_newline = part["SOURCE_FINAL_NEWLINE"]
-        if (
-            part["TOTAL_SOURCE_LINES"] != total_lines
-            or part["SOURCE_FINAL_NEWLINE"] != final_newline
-        ):
-            raise ValueError(f"{view_path}: inconsistent source metadata")
-        expected_next = int(part["LAST_SOURCE_LINE"]) + 1
-        parts.append(part)
-    if total_lines is None or expected_next != int(total_lines) + 1:
-        raise ValueError(f"{target}: incomplete source range")
-    numbered_rows = [row for part in parts for row in part["rows"]]
-    numbers = [number for number, _ in numbered_rows]
-    if numbers != list(range(1, int(total_lines) + 1)):
-        raise ValueError(f"{target}: global line range")
-    reconstructed = b"".join(
-        payload
-        + (b"\n" if number < int(total_lines) or bool(final_newline) else b"")
-        for number, payload in numbered_rows
-    )
-    actual = actual_override if actual_override is not None else (ROOT / target).read_bytes()
-    if reconstructed != actual:
-        raise ValueError(f"{target}: reconstructed bytes")
-    if sha256(actual).hexdigest() != expected_source_sha:
-        raise ValueError(f"{target}: literal source hash")
-    return {
-        "target_source": target,
-        "view_paths": view_paths,
-        "part_count": len(parts),
-        "total_source_lines": total_lines,
-        "source_final_newline": final_newline,
-        "source_bytes": len(actual),
-        "source_sha256": sha256(actual).hexdigest(),
-        "line_range": (numbers[0], numbers[-1]),
-        "contiguous": len(numbers) == len(set(numbers)) == int(total_lines),
-        "byte_identical": reconstructed == actual,
-    }
-
-
-def source_view_certificate(
-    target: str, view_paths: tuple[str, ...]
-) -> dict[str, object]:
-    certificate = validate_source_view_set(target, view_paths)
-    first_path = view_paths[0]
-    original = (ROOT / first_path).read_bytes()
-
-    def rejected(raw: bytes, *, enforce_view_hashes: bool) -> bool:
-        try:
-            validate_source_view_set(
-                target,
-                view_paths,
-                raw_overrides={first_path: raw},
-                enforce_view_hashes=enforce_view_hashes,
-            )
-        except (KeyError, SyntaxError, UnicodeDecodeError, ValueError):
-            return True
-        return False
-
-    prefix_mutation = original.replace(b"# C873SRC ", b"# C873SRX ", 1)
-    ordinal_mutation = original.replace(
-        b"PART_ORDINAL = 1\n", b"PART_ORDINAL = 9\n", 1
-    )
-    first_payload_end = original.index(b"\n", original.index(b"# C873SRC "))
-    payload_mutation = original[:first_payload_end] + b"X" + original[first_payload_end:]
-    try:
-        validate_source_view_set(
-            target,
-            view_paths,
-            actual_override=(ROOT / target).read_bytes() + b" ",
-        )
-    except ValueError:
-        child_mutation_detected = True
-    else:
-        child_mutation_detected = False
-    certificate["mutation_controls"] = {
-        "hard_pin_view_mutation_detected": rejected(
-            payload_mutation, enforce_view_hashes=True
-        ),
-        "malformed_prefix_detected": rejected(
-            prefix_mutation, enforce_view_hashes=False
-        ),
-        "ordinal_mutation_detected": rejected(
-            ordinal_mutation, enforce_view_hashes=False
-        ),
-        "payload_mutation_detected": rejected(
-            payload_mutation, enforce_view_hashes=False
-        ),
-        "actual_child_mutation_detected": child_mutation_detected,
-    }
-    return certificate
 
 
 def package_manifest() -> tuple[str, ...]:
@@ -585,7 +426,9 @@ def build_report() -> tuple[dict, list[str]]:
     source_views: dict[str, object] = {}
     for target, view_paths in SOURCE_VIEW_SETS.items():
         try:
-            certificate = source_view_certificate(target, view_paths)
+            certificate = source_view_certificate(
+                ROOT, target, view_paths, EXPECTED_ARTIFACT_SHA256
+            )
         except (KeyError, SyntaxError, UnicodeDecodeError, ValueError) as error:
             failures.append(f"source view: {target}: {error}")
         else:
